Finished with version 1 of the player list and player form views and functionality.
